+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Unreleased
 ----------
 
 * Add sunburst visualization for decision tree regressors
-* Add decisiont tree regression pipeline for subgroup discovery
+* Add decision tree regression pipeline for subgroup discovery
 
 2.0.0 (March 27, 2022)
 ----------------------